+++ demo/app/resource-select/resource-select-demo.component.ts
@@ -1,0 +1,15 @@
+import { SdkService } from './../../../packages/data/src/sdk/sdk.service';
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'ec-resource-select-demo',
+    templateUrl: 'resource-select-demo.component.html'
+})
+
+export class ResourceSelectDemoComponent implements OnInit {
+    multi;
+    solo;
+    constructor(public sdk: SdkService) { }
+
+    ngOnInit() { }
+}