+++ packages/core/src/field/field.ts
@@ -33,6 +33,12 @@ export class Field implements FieldConfigProperty {
   values: any[];
   /** Class string */
   class: string;
+  /** Pop Class(es) of eventually nested pops */
+  nestedPopClass: string;
+  /** If true, eventual pops will be active */
+  nestedPopActive: boolean;
+  /** If true, eventual pops will be visible*/
+  nestedPopVisible: boolean;
   /** wildcard for custom config values */
   [key: string]: any;